changes in the main form
how to proceed now?
create many forms, or use the same form then change the pictures and
itens only?
but i believe that the creation of many forms is better.
sorry about the mess in the github, but i could not revert some changes
that i didnt wanted to push.