changes in the changed gui
i remember that we should have only one main form, and the other
dialogs. so i recreated the new start window as a mainform and changed
the other mainform to dialog, recreating it as well as a room, the begin
room.
do not know the reason that the main window is executed by show, but
that is the way it works, otherwise the program would not close the
process properly after exit.
the file startwindow will be excluded at my next commit
the file mainwindow will be excluded a next commit

IMPORTANT: do not add includes into the cpp file. everything you need
you should include in the .h file